group blackbox tests by directory
